Added a library
Recoded the existsing groove.py to be more like a library.
The library contains everything you should need to get started.

The old wget download method has been booted in favor of an integrated
python downloader that supports callBacks on every blocks downloaded
(mainly for status updates to the user)

The functions has been rewritten to reuse as much code as possible,
while still keeping them simple.

Everything is the library is kept public, as the script still gets
banned, and the user might want to code his own procedure to download.

Cookies (session ID's) are now saved on exit, and reused if they are not
expired, possibly making grooveshark happier. (this add a new
dependency, cookielib, should be included with python)

QueueID's are now generated correctly

I'm using urlgrabber.keepalive to support HTTP/1.1 and keep-alive, this
is not included in the newer versions, so it requires an old version of
urlgrabber.

All prints have been removed, that is now left to the end user.